Change loss

+++ prepare.py
@@ -7,41 +7,53 @@ import pickle
 import logging
 
 
-def load_data(DATA):
+def load_data(DATA, output_type='reg'):
     i = {}
     DATA_PATH = Path('data') / DATA
     df = pd.read_csv(DATA_PATH / 'data.csv')
-    df['user'] = np.unique(df['user'], return_inverse=True)[1]  # Preprocess 0..N - 1
-    df['item'] = np.unique(df['item'], return_inverse=True)[1]
-    df['shifted_item'] = df['item'] + df['user'].nunique()
     i['trainval'] = pd.read_csv(DATA_PATH / 'trainval.csv')['index'].tolist()
     i['test'] = pd.read_csv(DATA_PATH / 'test.csv')['index'].tolist()
-    df_trainval = df.loc[i['trainval'], ['user', 'shifted_item', 'rating']]
-    df_test = df.loc[i['test'], ['user', 'shifted_item', 'rating']]
+    outcome_column = 'rating' if output_type == 'reg' else 'outcome'
+    df_trainval = df.loc[i['trainval'], ['user', 'shifted_item', outcome_column]]
+    df_test = df.loc[i['test'], ['user', 'shifted_item', outcome_column]]
     X_train = df_trainval[['user', 'shifted_item']].to_numpy()
-    y_train = df_trainval['rating'].to_numpy()
+    y_train = df_trainval[outcome_column].to_numpy()
     X_test = df_test[['user', 'shifted_item']].to_numpy()
-    y_test = df_test['rating'].to_numpy()
+    y_test = df_test[outcome_column].to_numpy()
+
+    '''for dataset in ['trainval', 'test']:
+                    if not (DATA_PATH / f'{DATA}.{dataset}_libfm').is_file():
+                        with open(DATA_PATH / f'{DATA}.{dataset}_libfm', 'w') as f:
+                            for user, item, outcome in np.array(df.loc[i[dataset], ['user', 'shifted_item', outcome_column]]):
+                                f.write('{:d} {:d}:1 {:d}:1\n'.format(outcome, user, item))'''
+
     return df['user'].nunique(), df['item'].nunique(), X_train, X_test, y_train, y_test
 
 def prepare_data(DATA):
     i = {}
     DATA_PATH = Path('data') / DATA
     df = pd.read_csv(DATA_PATH / 'data.csv')
+    # TODO Should detect if user and item are already properly sorted
     df['user'] = np.unique(df['user'], return_inverse=True)[1]  # Preprocess 0..N - 1
     df['item'] = np.unique(df['item'], return_inverse=True)[1]
     df['shifted_item'] = df['item'] + df['user'].nunique()
     # print(df.nunique(), df.min(), df.max())
-    print(df.agg(['min', 'max', 'nunique']))
+    # print(df.agg(['min', 'max', 'nunique']))
     i['trainval'] = pd.read_csv(DATA_PATH / 'trainval.csv')['index'].tolist()
     logging.warning('min trainval index %d max index %d', min(i['trainval']), max(i['trainval']))
     i['test'] = pd.read_csv(DATA_PATH / 'test.csv')['index'].tolist()
     logging.warning('min test index %d max index %d', min(i['test']), max(i['test']))
 
+    df['outcome'] = df['rating'].map(lambda value: int(value >= 4))
+    df.to_csv(DATA_PATH / 'data.csv', index=False)
+
     for dataset in ['trainval', 'test']:
-        with open(DATA_PATH / f'{DATA}.{dataset}_libfm', 'w') as f:
-            for user, item, rating in np.array(df.loc[i[dataset], ['user', 'shifted_item', 'rating']]):
-                f.write('{:d} {:d}:1 {:d}:1\n'.format(rating, user, item))
+        if not (DATA_PATH / f'{DATA}.{dataset}_libfm').is_file():
+            with open(DATA_PATH / f'{DATA}.{dataset}_libfm', 'w') as f:
+                for user, item, rating in np.array(df.loc[i[dataset], ['user', 'shifted_item', 'rating']]):
+                    f.write('{:d} {:d}:1 {:d}:1\n'.format(rating, user, item))
+        else:
+            logging.warning(f'Already existing OVBFM {dataset}')
 
 
 def prepare_ml_latest():
